Trinity.Core: extension capability enhancements;
